fix(tui): route actions to the correct component
The implemented structure passes Actions down to the next component when
ignored. First, a received Event or Action is handed to the currently
rendered popup (if one exists), and when an ActionResult::Ignored is
returned, the Event or Action is handed to either the Search or Page
component, depending on the context.

This fixes the issue where some Actions would be ignored due to them not
being routed to the correct component.