Move documentation to the docs subfolder
This moves the documentation site completely to the `docs/` subfolder.
Previously, the `mkdocs.yml` configuration file was at the root, but
with the change, it now lives under in the `docs/` folder.

The change also updates the versioning guidelines in the contributing
section of the documentation. It highlights how to deploy a new version