Remove GitHub Codespaces configuration
This configuration has been sitting in the repo unused for now 2 years.
And since we don't want to have unneeded files in the repo, this
configuration will be removed.

Do note that one can still work on wiki-tui in GitHub Codespaces without
this configuration, they just have to configure the container themselves