Add and implement a central Theme struct
This commit adds and implements a central Theme struct (living in
`config::Theme`). It is a first step for implementing configurable
theming inside of the application.

The Theme includes a set of standard colors for the various parts of the
application (foreground, background, selected background, etc.) and gets
passed down to every Component in the app, which will use the theme for
rendering.

It also includes two helper functions for generating a themed Block and
Paragraph, reducing duplicated code.

When we want to add more colors to the theme in the future, we just have
to add them to the struct and are able to use them in our Components