Add loading of a UserTheme through a config file
This implements the configuration of the Theme through a configuration
file. It lives under `WIKI_TUI_CONFIG/theme.toml`.

The structure of the loading is as follows: First, the default theme is
loaded and then, the UserTheme is loaded and only the values set in this
UserTheme are applied to the theme