Add BorderType configuration and user error message
This will add a configuration option for the BorderType of the borders
in the application. The option can be changed within the theme config.

It will also add a popup that appears when an error has occurred while
trying to load the theme configuration.

This change will also create an empty `theme.toml` file when there isn't
already an existing one