Implement search language changing via cli args
This will add the cli arg `-l <LANGUAGE>` and `--language <LANGUAGE>`
which will, when set, change the search language at startup to the
respective language.

For the language changing, the flag uses the builtin
`Language::From<String>` parser in order to allow using the language
code, the language name in english and the native language name to set
this value.

It also removes the debug cli command on release builds, which was never
intended to happen