Implement logging level changing via cli args
This will add the cli arg `--level` which will override the default
logging level, aswell as the env logging level.

The values have changed from the previous implementation. Now you need
to give it a value between 0 and 5, or the level in written form (eg.
"warn")

An improvement from the previous implementation is that we now have a
better and more verbose error message, when trying to set an invalid
logging level (previously, we just quietly failed, logging the error)