Implement printing config paths via cli args
This will add three cli args to print various config paths to the
console. After printing, the program will exit without launching the ui.

The three args are `--cache-dir`, `--config-path` and
`--theme-config-path` and are used to print their respective paths.

It also adds a simple help line to the args, which can be seen when
using `-h` or `--help`