Implement table of contents configuration options
This commit establishes a structure for load the configuration from the
config file and implements the configuration of the table of contents.

It deprecates and changes several configuration options for the table of
contents and updates the documentation site accordingly. Configuration
options for the page are now located inside of the `[page]` toml table
with the options for the table of contents being located at
`[page.toc]`.

With this commit, the application will now also distribute a pointer (`Arc` in
this case) to the loaded configuration to all components.

Signed-off-by: Builditluc <git@builditluc.eu>