Fix scrolling issues
This commit removes a recursion inside of `select_node` where a call to
`scroll_to_y` would potentially call itself with a different `y`
position due to it updating the selection, causing weird scrolling
behavior.

Now, only an update to the selection updates the viewport and an update
to the viewport only updates the selection.

Signed-off-by: Builditluc <git@builditluc.eu>