Fix compilation error with package managers
This commit will fix the compilation errors appearing when trying to
build wiki-tui for a package manager.

The underlying error is with the hard-coded path to the languages.json
in the proc-macro, which gets interpreted in multiple ways depending on
how wiki-tui is compiled (for example, `cargo publish` works, but
building through nixpkgs doesn't).

It fixes the issue by removing the path dependency completely and
instead passing the entire json file as a string literal to the
proc-macro

Signed-off-by: Builditluc <git@builditluc.eu>